% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_uncertainty_quotashare_MAIN.R
\name{calculate_landingshare}
\alias{calculate_landingshare}
\title{Calculate proportional fleet share of stock landings}
\usage{
calculate_landingshare(fleets, datayear, nyrs)
}
\arguments{
\item{fleets}{An \code{FLFleets} object}

\item{datayear}{The final year of data in the \code{FLFleets} object}

\item{nyrs}{The number of recent data years to include in calculations}
}
\value{
A data frame with variables: 'year', 'stock', 'fleet' and 'landingshare'
}
\description{
Function takes an \code{FLFleets} object and calculates the annual share of
overall stock landings for each fleet. It is assumed that there is no
stochasticity in landings estimates, and values are drawn from the first
iteration of each landings \code{FLQuant}.
}
