% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkCatchability}
\alias{checkCatchability}
\title{Simple function to check availability of catchability data}
\usage{
checkCatchability(logqs, qs_years, verbose, makeLog)
}
\arguments{
\item{qs_years}{A vector of years corresponding to catchability observations}

\item{verbose}{description}

\item{makeLog}{description}

\item{qs}{A matrix of time-series observations for catchability data where
cols = years and rows = stocks}

\item{makePlots}{description}
}
\value{
A list containing:

        - 'sdr': a TMB report summary for the optimised model parameters,
        - 'opt': a summary returned by the optimisation function,
        - 'obj': the optimised TMB objective function
}
\description{
Function takes a matrix of catchability time-series observations and checks
whether there are sufficient observations to fit a time-series state-space
model.
}
