% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{checkFail}
\alias{checkFail}
\title{Check for model fitting failure}
\usage{
checkFail(x, verbose)
}
\arguments{
\item{x}{a list}

\item{verbose}{(logical) should progress/summaries be printed? Defaults to
\code{TRUE}}
}
\description{
Check for failure. Return code.
}
