% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{findTrueZeros}
\alias{findTrueZeros}
\title{Crude classification of 'true' and 'false' zeros in historic data}
\usage{
findTrueZeros(dat)
}
\arguments{
\item{dat}{A matrix of historic effortshare where rows = years and cols = metiers}
}
\description{
Function takes a time-series of metier effort-share data as a matrix and
crudely classified any zero values as 'true' or 'false' zeros. A true zero
implies that an observation of zero effort-share is likely given historic
non-zero trends. A false zero implies that an observation of zero effort-share
is unlikely and may be due to missing data in that year. False zeros are
converted to \code{NA}.
}
