% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/conditioning_uncertainty_methods_wrapper_catchability.R
\name{fit_N_AR1}
\alias{fit_N_AR1}
\title{Wrapper function to fit latent AR1 with univariate normal intervals}
\usage{
fit_N_AR1(
  qs,
  metier_mt,
  qs_years,
  fillyear,
  deterministic,
  detMethod,
  verbose,
  makeLog,
  makePlots
)
}
\arguments{
\item{qs}{A matrix of time-series observations for catchability data where
cols = years and rows = stocks}

\item{metier_mt}{description}

\item{qs_years}{description}

\item{fillyear}{description}

\item{deterministic}{description}

\item{detMethod}{description}

\item{verbose}{description}

\item{makeLog}{description}

\item{makePlots}{description}
}
\value{
A list containing:

        - 'sdr': a TMB report summary for the optimised model parameters,
        - 'opt': a summary returned by the optimisation function,
        - 'obj': the optimised TMB objective function
}
\description{
A wrapper function to construct data and parameter inputs to build a TMB
objective function.
}
