% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/conditioning_uncertainty_methods_wrapper_catchability.R
\name{fitMVN_AR1}
\alias{fitMVN_AR1}
\title{Wrapper function to fit latent AR1 with with multivariate normal intervals}
\usage{
fitMVN_AR1(qs, metier_mt, qs_years, fillyear, verbose, makeLog)
}
\arguments{
\item{qs}{A matrix of time-series observations for catchability data where
cols = years and rows = stocks}

\item{metier_mt}{description}

\item{qs_years}{description}

\item{fillyear}{description}

\item{verbose}{description}

\item{makeLog}{description}
}
\value{
A list containing:

        - 'res' = ...
        - 'logs' = ...
        - 'plots' = ...
}
\description{
A wrapper function to construct data and parameter inputs to build a TMB
objective function.
}
