% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/conditioning_uncertainty_methods_wrapper_effortshare.R
\name{fitNhurdle_fixARrho}
\alias{fitNhurdle_fixARrho}
\title{Wrapper function to fit latent AR1 Gaussian with fixed AR1 correlation and hurdle Dirichlet and Bernoulli observation error}
\usage{
fitNhurdle_fixARrho(dat, rw, logitARrho)
}
\arguments{
\item{dat}{A matrix of time-series observations for compositional data where
cols = years and rows = compositional elements}

\item{rw}{A matrix of latent-state parameters to be optimised}
}
\value{
A list containing:

        - 'sdr': a TMB report summary for the optimised model parameters,
        - 'opt': a summary returned by the optimisation function,
        - 'obj': the optimised TMB objective function
}
\description{
A wrapper function to construct data and parameter inputs to build a TMB
objective function.
}
