% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/conditioning_uncertainty_methods_wrapper_effortshare.R
\name{fitNhurdle_fixb0b1}
\alias{fitNhurdle_fixb0b1}
\title{Wrapper function to fit latent AR1 Gaussian with hurdle Dirichlet observation error and fixed Bernoulli parameters}
\usage{
fitNhurdle_fixb0b1(dat, rw, logb0, logb1)
}
\arguments{
\item{dat}{A matrix of time-series observations for compositional data where
cols = years and rows = compositional elements}

\item{rw}{A matrix of latent-state parameters to be optimised}
}
\value{
A list containing:

        - 'sdr': a TMB report summary for the optimised model parameters,
        - 'opt': a summary returned by the optimisation function,
        - 'obj': the optimised TMB objective function
}
\description{
A wrapper function to construct data and parameter inputs to build a TMB
objective function.
}
