% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_forecast.R
\name{forecast_from_AR1_logN}
\alias{forecast_from_AR1_logN}
\title{Stochastic forecast from an AR1 latent process with log-normal observation}
\usage{
forecast_from_AR1_logN(dat, pl, plsd, MVNcov = NULL, fillyear, nit)
}
\description{
...
}
