% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_forecast.R
\name{forecast_from_hurdle}
\alias{forecast_from_hurdle}
\title{Stochastic forecast from an AR1 latent process with Dirichlet observation}
\usage{
forecast_from_hurdle(dat, pl, plsd, MVNcov, fillyear, nit)
}
\description{
...
}
