% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_forecast.R
\name{forecast_from_rw}
\alias{forecast_from_rw}
\title{Stochastic forecast from a random walk model with Dirichlet observation}
\usage{
forecast_from_rw(dat, pl, plsd, MVNcov, fillyear, nit)
}
\description{
...
}
