% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{impute_cases}
\alias{impute_cases}
\title{Small function to impute values to cases of zero and one in proportional data}
\usage{
impute_cases(dat)
}
\arguments{
\item{dat}{A data frame containing numeric data.}
}
\description{
Function identifies years where a metier has zero effort.
In these cases, impute 1e-6 to the proportional
effort-share of each zero effort metier. Then deduct the
added effort-share from non-zero effort metiers.
}
