% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_forecast.R
\name{insert_deterministic}
\alias{insert_deterministic}
\title{Insert results of a deterministic calculation into FLR}
\usage{
insert_deterministic(qs0, qs_years, metier_mt, deterministic_yrs, nit)
}
\description{
In the case of model failure or insufficient data to fit a model, carry out
a deterministic calculation based on the mean of the historic period
}
