% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_forecast.R
\name{insert_forecast}
\alias{insert_forecast}
\title{Insert results of stochastic forecast into FLR}
\usage{
insert_forecast(
  resvariates,
  qs0,
  qs_years,
  metier_mt,
  deterministic,
  deterministic_yrs,
  detMethod,
  nit
)
}
\description{
...
}
