% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{invlogit}
\alias{invlogit}
\title{Multinomial inverse logit transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{a numeric vector on a natural scale}
}
\description{
Multinomial inverse logit transformation of a vector.
}
