% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_fit_forecast_ggplot.R
\name{plot_fit_Dir}
\alias{plot_fit_Dir}
\title{Plot latent-process model fit for a random walk/AR1 and Dirichlet noise state-space model}
\usage{
plot_fit_Dir(data, pl, plsd, years, invlogitfun)
}
\arguments{
\item{data}{named list input to TMB containing observation data}

\item{pl}{named list output from TMB containing parameter estimates}

\item{plsd}{named list output from TMB containing parameter standard error}

\item{years}{vector of years associated with observations and forecasts}

\item{invlogitfun}{function to perform multinomial inverse logit transformation
of latent process values.}
}
\description{
Plot the marginal maximum likelihood estimate for the latent process of a
random walk (or AR1) state-space model with Dirichlet observation noise.
The 95% confidence intervals around the estimate are shown as dashed lines.
Observation data are shown as open points.
}
