% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_fit_forecast_ggplot.R
\name{plot_fit_MVN}
\alias{plot_fit_MVN}
\title{Plot latent-process model fit for a random walk and noise state-space model}
\usage{
plot_fit_MVN(data, pl, plsd, years)
}
\arguments{
\item{data}{named list input to TMB containing observation data}

\item{pl}{named list output from TMB containing parameter estimates}

\item{plsd}{named list output from TMB containing parameter standard error}

\item{years}{vector of years associated with observation data}
}
\description{
Plot the marginal maximum likelihood estimate for the latent process of a
random walk and observation noise state-space model together with 95%
confidence intervals around the estimate and observation points.
}
