% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_fit_forecast_ggplot.R
\name{plot_forecast_Dir}
\alias{plot_forecast_Dir}
\title{Plot latent-process model fit and forecast for a random walk/AR1 and Dirichlet noise state-space model}
\usage{
plot_forecast_Dir(data, pl, plsd, pred_quantiles, years, invlogitfun)
}
\arguments{
\item{data}{named list input to TMB containing observation data}

\item{pl}{named list output from TMB containing parameter estimates}

\item{plsd}{named list output from TMB containing parameter standard error}

\item{pred_quantiles}{array of forecast value summaries at the 2.5, 50, 97.5
centiles.}

\item{years}{vector of years associated with observations and forecasts}

\item{invlogitfun}{function to perform multinomial inverse logit transformation
of latent process values.}
}
\description{
Plot the marginal maximum likelihood estimate and median stochastic forecast
for the latent process of a random walk (or AR1) state-space model with Dirichlet
observation noise. The 95% confidence intervals around the estimate and forecast
are shown as dashed lines. Observation data are shown as open points.
}
