% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Simple function to draw random Dirichlet distributed draws}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of draws}

\item{alpha}{a numeric vector of alpha parameters.}
}
\description{
Function takes a vector of alpha parameters and samples $n$ random draws
from a Dirichlet distribution.
}
