% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_uncertainty_methods_TMB_logMVNrw.R
\name{TMB_logMVNrw}
\alias{TMB_logMVNrw}
\title{Forecast catchability from fitted multivariate log-normal random walk}
\usage{
TMB_logMVNrw(
  qs,
  metier_mt,
  fillyear,
  deterministic,
  deterministic_yrs,
  detMethod,
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE
)
}
\arguments{
\item{qs}{A matrix of historic catchability where rows = years and cols = stocks}

\item{fillyear}{(Integer) Vector of years to be filled with forecasted catchability}

\item{deterministic}{(Logical) Should the results for first iteration be
a simple mean of the historical period? Defaults
to \code{TRUE}}

\item{detMethod}{(Character) The method to calculate the deterministic average.
If "mean" then a simple mean of the most recent three data years
is used. Otherwise, the median of the stochastic forecast is
used.}

\item{verbose}{(Logical) Should the function print progress? Defaults
to \code{TRUE}}

\item{makeLog}{(Logical) Should the function generate a log of the methods
applied to each stock (and their success)? Defaults
to \code{TRUE}}

\item{makePlots}{(Logical) Should the function generate a figure of model fit?
Defaults to \code{TRUE}}

\item{meter_mt}{An object of class \code{FLMetier}, \code{FLMetier},
\code{FLMetierExt} or a \code{FLMetierExt}.}
}
\value{
An updated FLR metier object.
}
\description{
Function takes a time-series of stock catchability data as a matrix and
fits a multivariate normal random walk to the time-series. The final data
year is simulated to generate a distribution based on observation uncertainty.
A forecast is carried out for each simulated catchability value that propagates
uncertainty estimated from the random walk.
}
\details{
A useful feature of this method is that missing values are easily accommodated.
}
