% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_uncertainty_catchability_MAIN.R
\name{uncertainty_catchability}
\alias{uncertainty_catchability}
\alias{uncertainty_catchability,FLFleetExt-method}
\alias{uncertainty_catchability,FLFleetsExt-method}
\alias{uncertainty_catchability,FLFleet-method}
\alias{uncertainty_catchability,FLFleets-method}
\title{Stochastically sample future catchability from historic estimates}
\usage{
uncertainty_catchability(
  fleets,
  method = "TMB_logMVNrw",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  parallel = FALSE
)

\S4method{uncertainty_catchability}{FLFleetExt}(
  fleets,
  method = "TMB_logMVNrw",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE
)

\S4method{uncertainty_catchability}{FLFleetsExt}(
  fleets,
  method = "TMB_logMVNrw",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  parallel = FALSE
)

\S4method{uncertainty_catchability}{FLFleet}(
  fleets,
  method = "TMB_logMVNrw",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE
)

\S4method{uncertainty_catchability}{FLFleets}(
  fleets,
  method = "TMB_logMVNrw",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{fleets}{An object of class \code{FLFleet}, \code{FLFleets},
\code{FLFleetExt} or a \code{FLFleetsExt}.}

\item{method}{(Character) The algorithm used to fit the statistical model and
generate stochastic forecasts. Using the default is highly
recommended.}

\item{datayears}{A vector of historic years used to estimate future fleet
catchability}

\item{TACyear}{(Integer) The projection year in the fleets object}

\item{deterministic}{(Logical) Should the results for first iteration be
a deterministic average of the historical period? Defaults
to \code{TRUE}}

\item{deterministic_yrs}{(Integer) The number of recent data years to use in
the deterministic calculation. Defaults to 3.}

\item{detMethod}{(Character) The method to calculate the deterministic average.
If "mean" then a simple mean of the most recent data years
is used. Otherwise, the median of the stochastic forecast is
used.}

\item{verbose}{(Logical) Should the function print progress? Defaults
to \code{TRUE}}

\item{makeLog}{(Logical) Should the function generate a log of the methods
applied to each stock (and their success)? Defaults
to \code{TRUE}}

\item{makePlots}{(Logical) Should the function generate plots showing model
fit to data for each stock? Defaults to \code{TRUE}}
}
\value{
named list
}
\description{
Function takes a \code{FLFleet}, \code{FLFleets}, \code{FLFleetExt} or a
\code{FLFleetsExt} object and fits a statistical model to historic
observations of metier-stock catchability for each fleet. The catchability
for future years is a stochastic forecast from fitted model parameters.
}
