% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_uncertainty_effortshare_MAIN.R
\name{uncertainty_effortshare}
\alias{uncertainty_effortshare}
\alias{uncertainty_effortshare,FLFleetExt-method}
\alias{uncertainty_effortshare,FLFleetsExt-method}
\alias{uncertainty_effortshare,FLFleet-method}
\alias{uncertainty_effortshare,FLFleets-method}
\title{Stochastically forecast future metier effort-share from historic estimates}
\usage{
uncertainty_effortshare(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  removeNonRecent = FALSE
)

\S4method{uncertainty_effortshare}{FLFleetExt}(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  removeNonRecent = FALSE
)

\S4method{uncertainty_effortshare}{FLFleetsExt}(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  removeNonRecent = FALSE
)

\S4method{uncertainty_effortshare}{FLFleet}(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  removeNonRecent = FALSE
)

\S4method{uncertainty_effortshare}{FLFleets}(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  datayears = NULL,
  TACyear = NULL,
  deterministic = TRUE,
  deterministic_yrs = 3,
  detMethod = "mean",
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE,
  removeNonRecent = FALSE
)
}
\arguments{
\item{fleets}{An object of class \code{FLFleet}, \code{FLFleets},
\code{FLFleetExt} or a \code{FLFleetsExt}.}

\item{method}{(Character) The algorithm used to fit the statistical model and
generate stochastic forecasts. Using the default is highly
recommended.}

\item{datayears}{A vector of historic years used to estimate future fleet
effort-share}

\item{TACyear}{(Integer) The projection year in the fleets object}

\item{deterministic}{(Logical) Should the results for first iteration be
a simple mean of the historical period? Defaults
to \code{TRUE}}

\item{deterministic_yrs}{(Integer) The number of recent data years to use in
the deterministic calculation. Defaults to 3.}

\item{detMethod}{(Character) The method to calculate the deterministic average.
If "mean" then a simple mean of the most recent data years
is used. Otherwise, the median of the stochastic forecast is
used.}

\item{verbose}{(Logical) Should the function print progress? Defaults
to \code{TRUE}}

\item{makeLog}{(Logical) Should the function record model fitting and forecasting
performance for each fleet? Defaults to \code{TRUE}}

\item{makePlots}{(Logical) Should the function generate a figure showing
observations and model fit and forecasts with uncertainty
intervals for each fleet? Defaults to \code{TRUE}}

\item{removeNonRecent}{(Logical) Should the function remove fleets that have not
registered landings in recent years? Defaults to \code{FALSE}}
}
\value{
a named list containing:
        - **fleets** - An updated \code{FLFleet}, \code{FLFleets},
              \code{FLFleetExt} or \code{FLFleetsExt} object
        - **logs** - An optional summary of model fitting and forecasting
                     performance for each fleets
        - **plots** - An optional figure showing the model maximum likelihood
                      fit (+ 95% CI) and median (+ 95% CI) forecasts against
                      observations for each fleet.
}
\description{
Function takes a \code{FLFleet}, \code{FLFleets}, \code{FLFleetExt} or a
\code{FLFleetsExt} object and fits a statistical model to historic
observations of metier effort-share for each fleet. The metier effort-share
for future years is a stochastic forecast from fitted model parameters.
}
