% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning_uncertainty_quotashare_MAIN.R
\name{uncertainty_quotashare}
\alias{uncertainty_quotashare}
\alias{uncertainty_quotashare,FLFleetsExt-method}
\alias{uncertainty_quotashare,FLFleets-method}
\title{Stochastically sample future fleet quota-share from historic estimates}
\usage{
uncertainty_quotashare(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  quotashare = NULL,
  datayear = NULL,
  TACyear = NULL,
  nyrs = 4,
  deterministic = TRUE,
  deterministic_yrs = 3,
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE
)

\S4method{uncertainty_quotashare}{FLFleetsExt}(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  quotashare = NULL,
  datayear = NULL,
  TACyear = NULL,
  nyrs = 4,
  deterministic = TRUE,
  deterministic_yrs = 3,
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE
)

\S4method{uncertainty_quotashare}{FLFleets}(
  fleets,
  method = "TMB_DirMissingAR1Hurdle",
  quotashare = NULL,
  datayear = NULL,
  TACyear = NULL,
  nyrs = 4,
  deterministic = TRUE,
  deterministic_yrs = 3,
  verbose = TRUE,
  makeLog = TRUE,
  makePlots = TRUE
)
}
\arguments{
\item{fleets}{An object of class \code{FLFleet}, \code{FLFleets},
\code{FLFleetExt} or a \code{FLFleetsExt}.}

\item{method}{(Character) The algorithm used to fit the statistical model and
generate stochastic forecasts. Using the default is highly
recommended.}

\item{quotashare}{(Optional) A data frame containing: year, stock, fleet
and landingshare.}

\item{datayear}{(Integer) The final data year in the fleets object}

\item{TACyear}{(Integer) The projection year in the fleets object}

\item{nyrs}{(Integer) The number of historic years used to estimate
future fleet quota-share}

\item{deterministic}{(Logical) Should the results for first iteration be
a simple mean of the historical period? Defaults
to \code{TRUE}}

\item{deterministic_yrs}{(Integer) The number of recent data years to use in
the deterministic calculation. Defaults to 3.}

\item{verbose}{(Logical) Should the function print progress? Defaults
to \code{TRUE}}

\item{makeLog}{(Logical) Should the function record model fitting and forecasting
performance for each fleet? Defaults to \code{TRUE}}

\item{makePlots}{(Logical) Should the function generate a figure showing
observations and model fit and forecasts with uncertainty
intervals for each fleet? Defaults to \code{TRUE}}
}
\description{
Function takes a \code{FLFleet}, \code{FLFleets}, \code{FLFleetExt} or a
\code{FLFleetsExt} object and fits a statistical model to historic
observations of stock quota-share for each fleet. The stock quota-share
for future years is a stochastic forecast from fitted model parameters.
}
